[FIX] Objects box pixels size percent + correct object combobox issue

+++ Editor/Models/System/systemobject3d.cpp
@@ -25,8 +25,8 @@ SystemObject3D::SystemObject3D() :
 
 SystemObject3D::SystemObject3D(int i, QString n, ShapeKind shapeKind, int objID,
     int mtlID, int pictureID, ObjectCollisionKind collisionKind, int
-    collisionCustomID, double scale, int wS, int wP, int hS, int hP, int dS, int
-    dP, bool stretch) :
+    collisionCustomID, double scale, int wS, double wP, int hS, double hP, int
+    dS, double dP, bool stretch) :
     SystemSpecialElement(i, n, shapeKind, objID, mtlID, pictureID, collisionKind
         , collisionCustomID, scale, wS, wP, hS, hP, dS, dP, stretch)
 {
