+++ Editor/Models/System/systemobject3d.h
@@ -30,8 +30,8 @@ public:
     SystemObject3D(int i, QString n, ShapeKind shapeKind = ShapeKind::Box, int
         objID = -1, int mtlID = -1, int pictureID = -1, ObjectCollisionKind
         collisionKind = ObjectCollisionKind::None, int collisionCustomID = -1,
-        double scale = 1.0, int wS = 1, int wP = 0, int hS = 1, int hP = 0, int
-        dS = 1, int dP = 0, bool stretch = false);
+        double scale = 1.0, int wS = 1, double wP = 0, int hS = 1, double hP = 0
+        , int dS = 1, double dP = 0, bool stretch = false);
     virtual ~SystemObject3D();
 
     virtual SystemPicture* picture() const;
